requirements.txt
pydantic>=2.0
python-dotenv>=1.0
